# openapi.yaml
openapi: 3.0.0
info:
  title: 'Delilah Resto '
  description: API for delicious food orders. As a customer you can register, see the list of our products and place an order.Restaurant managers have the ability to receive orders and update them..
  termsOfService: https://github.com/hmfarias/03-acamica-delilah#readme
  contact:
    name: Marcelo Farias
    email: hmfarias7@gmail.com
  version: 1.0.0
servers:

  - description: Local Server
    url: http://localhost:3000/api/v1
...
paths:
...
  /users:
   # --------------------------------------------GET ALL USERS    
    get:
      tags:
        - Users
      security: 
        - bearerAuth: 
          - read      
      summary: 'List all users'
      description: 'Admin level required'
      operationId: getUsers
      responses:
        '200':
          description: 'Successful operation - returns an array of users'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/usersResp200'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'There are no users in the database'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'           
  /users/{id}:
    # --------------------------------------------GET USER BY ID
    get:
      tags:
        - Users
      security: 
        - bearerAuth: 
          - read
      summary: 'Find user by Id - Return a single user'
      description: 'Admin level or account owner required '
      operationId: getUser
      parameters:
        - name: id
          in: path
          description: 'ID of user to return'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the user with the Id specified by parameter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userResp200'      
        '403':
          description: 'Admin Level or account owner required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403-2'         
        '404':
          description: 'The user not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                   
        '410':
          description: 'The user is deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'                   
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'       
  /users/update/{id}:
    # -------------------------------------------PUT UPDATE USER          
    put:
      tags:
        - Users
      security: 
        - bearerAuth: 
          - read
      summary: 'Update an existent user'
      description: 'Admin level required, or be the account owner - You can update the [name], [email], [phone] and [password] fields - You can not update the [username] or [user role] fields'
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: 'ID of user that needs to be updated'
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: 'Required to update a user'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userReqUpdate'
        required: true
      responses:
        '200':
          description: 'Successful operation - returns the updated user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userResp200'          
        '400':
          description: 'Id parameter or some field is missing or has an incorrect format'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level or data owner required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403-2'                   
        '404':
          description: 'User not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'       
        '500':
          description: 'The user could not be updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                                   
        'default':
          description: 'Unexpected error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'       
  /users/delete/{id}:
      # ---------------------------------------DELETE USER BY ID
    delete:
      tags:
        - Users
      security: 
        - bearerAuth: 
          - read      
      summary: 'Deletes a user by Id'
      description: 'Admin level required'
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          description: 'ID of user that needs to be deleted'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the deleted user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userResp200'          
        '400':
          description: 'The Id parameter is required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The user not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                     
        '410':
          description: 'The user is already deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'     
        '500':
          description: 'The user could not be deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                           
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'    
  /users/restore/{id}:
    # ----------------------------------------RESTORE USER BY ID
    put:
      tags:
        - Users
      security: 
        - bearerAuth: 
          - read      
      summary: 'Restore a user by Id'
      description: 'Admin Level required'
      operationId: restoreUser
      parameters:
        - name: id
          in: path
          description: 'ID of user that needs to be restored'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Successful operation - returns the restored user'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userResp200'          
        '400':
          description: 'The Id parameter is required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error400Param'      
        '403':
          description: 'Admin Level required'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error403'                   
        '404':
          description: 'The user not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error404'                     
        '406':
          description: 'The user is not deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error406'                     
        '410':
          description: 'The user is deleted (soft deleted)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error410'     
        '500':
          description: 'The user could not be deleted'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error500'                           
        'default':
          description: 'Internal error - Try again later'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorDefault'        


  # ------------------------------------------------------------PATHS FOR PAYMENT METHODS 
...
components:
...
  securitySchemes:
    bearerAuth:            
      type: http
      scheme: bearer
      bearerFormat: JWT

  #  --------------------------------------------------------SCHEMAS       
  schemas:
...
    roleName:
      properties:
        name:
          type: string
          example: 'user'
...
    userResp:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 7
        username:
          type: string
          example: 'jSmith'
        name:
          type: string
          example: 'John Smith'
        email:
          type: string
          example: 'jsmith@gmail.com'
        phone:
          type: string
          example: '+44 8812345678'
        address:
          type: string
          example: 'TARDIST ST, 4242, Space 453, Gallifrey'
        createdAt:
          type: string
          example: '|creation date and time|'
        updatedAt:
          type: string
          example: '|update date and time|'
        deletedAt:
          type: string
          example: '|null| or |delete date and time|'
        role:
             $ref: '#/components/schemas/roleName'
    userReqUpdate:
      type: object
      properties:
        name:
          type: string
          example: 'John Smith'
        email:
          type: string
          example: 'jsmith@gmail.com'
        phone:
          type: string
          example: '+44 8812345678'
        address:
          type: string
          example: 'TARDIST ST, 4242, Space 453, Gallifrey'
        password:
          type: string
          example: '***************'
    userResp200:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/userResp'      
        message:
          $ref: '#/components/schemas/messageSuccess'
    usersResp200:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        data:
          type: object
          properties:
            users:
              $ref: '#/components/schemas/userResp'     
        message:
          type: string
          example: 'Successfully recovered Users'

      #  ----------------------------------SCHEMAS FOR PAYMENT METHODS
...
    messageSuccess:
      type: string
      example: 'Successful operation for |entity| ID: |entity ID|, Name: |entity name|'     
          
    #  ----------------------------------SCHEMAS FOR ERRORS
    errorDefault:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
          properties:
            error:
              type: string
              example: '|Error|'
        message:
          type: string
          example: 'Internal error - Try again later'
...
    error400Param:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|parameter or field| |parameter or field name| is required | or has an incorrect format'  
    error403:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: 'Admin level required'  
    error403-2:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: 'Admin level or data owner required'  
    error404:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| not found'     
    error406:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| is not deleted'     
...
    error410:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: '|entity| is deleted - (soft deleted)'
    error500:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        data:
          type: object
        message:
          type: string
          example: 'Unexpected error - |message|'